+++ src/dznpy/adv_shell/common.py
@@ -15,7 +15,8 @@ from .. import cpp_gen, ast
 from ..code_gen_common import BLANK_LINE, GeneratedContent
 from ..cpp_gen import Comment, Constructor, Function, MemberVariable, Fqn, Namespace, Struct, \
     TypeDesc
-from ..misc_utils import NameSpaceIds, TextBlock, plural
+from ..misc_utils import NameSpaceIds, TextBlock
+from ..scoping import NameSpaceIds
 
 # own modules
 from .types import RuntimeSemantics
@@ -184,7 +185,7 @@ class DznElements:
 @dataclass(frozen=True)
 class CppEncapsulee:
     """Data class comprising attributes of a C++ encapsulee."""
-    member_var: MemberVariable or None
+    member_var: MemberVariable
     name: str
 
     def __str__(self):