+++ test/unit_tests/test_misc_utils.py
@@ -13,37 +13,6 @@ from unittest import TestCase
 from dznpy.misc_utils import *
 
 
-def test_create_namespaceids_type():
-    correct: NameSpaceIds = ['My', 'Project']
-    assert is_namespaceids_instance(namespaceids_t(correct)), 'Pass through correct type'
-    assert namespaceids_t('My.Project') == correct
-    assert namespaceids_t('MyComponent') == ['MyComponent']
-    assert is_namespaceids_instance(namespaceids_t('MyComponent')) is True
-
-
-def test_check_and_assert_namespaceids_type():
-    correct: NameSpaceIds = ['My', 'Project']
-    assert_namespaceids_instance(correct)
-    assert is_namespaceids_instance(correct) is True
-    assert is_namespaceids_instance([]) is True, 'an empty list is ok'
-
-    exc_msg = 'The argument is not a NameSpaceIds type'
-    assert is_namespaceids_instance(None) is False
-    with pytest.raises(TypeError) as exc:
-        assert_namespaceids_instance(None)
-    assert str(exc.value) == exc_msg
-
-    assert is_namespaceids_instance('My.Project') is False
-    with pytest.raises(TypeError) as exc:
-        assert_namespaceids_instance('My.Project')
-    assert str(exc.value) == exc_msg
-
-    assert is_namespaceids_instance(['One', 2, 3]) is False
-    with pytest.raises(TypeError) as exc:
-        assert_namespaceids_instance(['One', 2, 3])
-    assert str(exc.value) == exc_msg
-
-
 def test_check_is_str_set():
     assert is_strset_instance({'My', 'Project'}) is True
     assert is_strset_instance(set()) is True
@@ -74,29 +43,6 @@ def test_flatten_to_strlist():
     assert flatten_to_strlist(['One', '', 'Two'], skip_empty_strings=False) == ['One', '', 'Two']
     assert flatten_to_strlist(['One', [''], 'X'], skip_empty_strings=False) == ['One', '', 'X']
     assert flatten_to_strlist([{123: '', 456: None}, ['Y']], skip_empty_strings=False) == ['', 'Y']
-
-
-def test_scope_resolution_order_ok():
-    assert scope_resolution_order(searchable=namespaceids_t('IToaster'),
-                                  calling_scope=namespaceids_t(['My', 'Project'])) == \
-           [['My', 'Project', 'IToaster'], ['My', 'IToaster'], ['IToaster']]
-
-    assert scope_resolution_order(searchable=namespaceids_t(['My', 'ILed']),
-                                  calling_scope=namespaceids_t(['My', 'Project'])) == \
-           [['My', 'Project', 'My', 'ILed'], ['My', 'My', 'ILed'], ['My', 'ILed']]
-
-    assert scope_resolution_order(searchable=namespaceids_t(['My', 'ILed']),
-                                  calling_scope=None) == [['My', 'ILed']]
-
-
-def test_scope_resolution_order_fail():
-    with pytest.raises(TypeError) as exc:
-        scope_resolution_order(searchable=None, calling_scope=namespaceids_t(['NS']))
-    assert str(exc.value) == 'The argument is not a NameSpaceIds type'
-
-    with pytest.raises(TypeError) as exc:
-        scope_resolution_order(searchable=123, calling_scope=namespaceids_t(['NS']))
-    assert str(exc.value) == 'The argument is not a NameSpaceIds type'
 
 
 def test_plural_ok():